Feat: 존재하는 파일 생략 기능 추가 #24

+++ parallelism/util/youtube_livechat_crawling_nonBuffer.py
@@ -1,3 +1,5 @@
+import os
+
 import pytchat
 import pandas as pd
 from collections import defaultdict
@@ -8,28 +10,32 @@ from parallelism.exception.LiveChatUnknownException import \
 
 
 async def live_chat(videoId, folder):
-    chat = pytchat.create(video_id=videoId)
-    file_path = f"./{folder}/{videoId}.csv"
-    data = defaultdict(list)
-    while chat.is_alive():
-        chatdata = chat.get()
-        for item in chatdata.items:
-            data["ID"].append(item.id)
-            data["Name"].append(item.author.name)
-            data["comment"].append(item.message)
-            data["Date"].append(item.datetime)
-            data["time"].append(item.elapsedTime)
+    filename = videoId+'.csv'
+    file_path = os.path.join(folder,filename)
+    if os.path.exists(filename):
+        print(f"파일 '{file_path}'이 이미 존재하여 넘어갑니다.")
+    else:
+        chat = pytchat.create(video_id=videoId)
+        data = defaultdict(list)
+        while chat.is_alive():
+            chatdata = chat.get()
+            for item in chatdata.items:
+                data["ID"].append(item.id)
+                data["Name"].append(item.author.name)
+                data["comment"].append(item.message)
+                data["Date"].append(item.datetime)
+                data["time"].append(item.elapsedTime)
 
-    result = pd.DataFrame(data)
-    result.to_csv(file_path, mode="a", index=False)
+        result = pd.DataFrame(data)
+        result.to_csv(file_path, mode="a", index=False)
 
-    try:
-        # 채팅이 더이상 존재하지 않은면 exception 발생
-        chat.raise_for_status()
-    except Exception as e:
-        if type(e).__name__ == "ChatDataFinished":
-            pass
-        elif type(e).__name__ == "NoContents":
-            raise EmptyLiveChatException()
-        else:
-            raise LiveChatUnknownException()
+        try:
+            # 채팅이 더이상 존재하지 않은면 exception 발생
+            chat.raise_for_status()
+        except Exception as e:
+            if type(e).__name__ == "ChatDataFinished":
+                pass
+            elif type(e).__name__ == "NoContents":
+                raise EmptyLiveChatException()
+            else:
+                raise LiveChatUnknownException()
